move test utilites out of /test
