+++ cmd/kubeadm/app/util/certs/BUILD
@@ -8,7 +8,7 @@ load(
 go_library(
     name = "go_default_library",
     srcs = ["util.go"],
-    importpath = "k8s.io/kubernetes/cmd/kubeadm/test/certs",
+    importpath = "k8s.io/kubernetes/cmd/kubeadm/app/util/certs",
     deps = [
         "//cmd/kubeadm/app/util/pkiutil:go_default_library",
         "//staging/src/k8s.io/client-go/util/cert:go_default_library",
