+++ cmd/kubeadm/test/kubeconfig/util.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
-	certstestutil "k8s.io/kubernetes/cmd/kubeadm/test/certs"
+	certstestutil "k8s.io/kubernetes/cmd/kubeadm/app/util/certs"
 )
 
 // AssertKubeConfigCurrentCluster is a utility function for kubeadm testing that asserts if the CurrentCluster in
